Создать класс первоначальной конфигурации приложения AppConfiguration и реализовать конфигурирование перечня сервисов для загрузки
