+++ ArtBot/Program.cs
@@ -1,7 +1,7 @@
-﻿using ArtBot.Services;
+﻿using ArtBot;
+using ArtBot.Services;
 using ArtBot.Services.BotServices;
 using ArtBot.Services.BotServices.TG;
-using System.Collections.Generic;
 
 
 /// <summary>
@@ -15,28 +15,19 @@ class Program
 {
     static async Task Main(string[] args)
     {
-        // Создание списка экземпляров ботов
-        var listBots = new List<IBotService>();
+        //Сконфигурировать параметры загрузки приложения
+        var appConfiguration = new Configuration();
 
-        listBots.AddRange(Enum.GetValues(typeof(BotType))
-                .Cast<BotType>()
-                .Select(botType => GetBotService(botType)));
-
-        // Создание списка экземпляров сервисов
-        var listServices = new List<IService>();
-
-        listServices.AddRange(Enum.GetValues(typeof(ServiceType))
-            .Cast<ServiceType>()
-            .Select(serviceType => GetService(serviceType, listBots)));
+        
 
 
 
 
 
-        var botService = new BotService();
+        //var botService = new BotService();
 
         // Запуск сервисов
-        await botService.StartAsync();
+        //await botService.StartAsync();
         /*
         //await responseService.StartAsync();
         //await messageHandlingService.StartAsync();
@@ -57,34 +48,7 @@ class Program
         */
     }
 
-    private static IBotService GetBotService(BotType botType)
-    {
-        return botType switch
-        {
-            BotType.Telegram => new TelegramBot(),
-          /*//BotType.VK => new VKBot(),
-            //BotType.Viber => new ViberBot(),
-            //BotType.WhatsApp => new WhatsAppBot(),*/
-            _ => throw new ArgumentOutOfRangeException(),
-        };
-    }
-
-    private static IService GetService(ServiceType serviceType, List<IBotService> listBots)
-    {
-        return serviceType switch
-        {
-            ServiceType.BotService => new BotService(listBots),
-          /*//case ServiceType.ResponseService: return new ResponseService();
-            //case ServiceType.MessageHandlingService: return new MessageHandlingService();
-            //case ServiceType.LoggingService: return new LoggingService();
-            //case ServiceType.DatabaseProcessor: return new DatabaseProcessor();
-            //case ServiceType.WebhookService: return new WebhookService();
-            //case ServiceType.AuthService: return new AuthService();
-            //case ServiceType.NotificationService: return new NotificationService();
-            //case ServiceType.RequestHandlerService: return new RequestHandlerService();*/
-            _ => throw new ArgumentOutOfRangeException(),
-        };
-    }
+    
 
     static async Task<string> GetRequestFromClientAsync()
     {
