+++ ArtBot/Services/Service.cs
@@ -4,14 +4,14 @@ namespace ArtBot.Services
 {
     /// <summary>
     /// Класс AppService, реализует интерфейс IService. 
-    /// Класс содержит поле List<IService> _listServices, которое используется для хранения списка сервисов. 
+    /// Класс содержит свойство List<IService> ListServices, которое используется для хранения списка сервисов. 
     /// Метод StartAsync запускает все сервисы приложения, используя метод Task.WhenAll. 
     /// Поле со списком сервисов инициализируется через входной параметр конструктора класса.
     /// </summary>
-    public class AppService(List<IBotService> listServices) : IService
+    public class Service(List<IBotService> listServices) : IService
     {
-        private readonly List<IBotService> _listServices = listServices;
+        public List<IBotService> ListServices { get; } = listServices;
 
-        public Task StartAsync() => Task.WhenAll(_listServices.Select(s => s.StartAsync()));
+        public Task StartAsync() => Task.WhenAll(ListServices.Select(s => s.StartAsync()));
     }
 }
